requirements.txt
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xarray==2023.6.0
netCDF4==1.6.4
cdsapi==0.6.1
python-dotenv==1.0.0
requests==2.31.0
